Add horizontal radar appx option
The traditional approach to multi-Doppler synthesis is to ignore
the w component of the Doppler velocity and solve for w via mass
continuity. For the airborne case, there are many steep angle
observations that can add information to the retrieval if the w
component is retained. This is the default mode in SAMURAI.
However, the resulting vertical velocity can be unstable due to
ill-conditioned normal equations just above the horizontal plane.
This option allows users to ignore the w component and perform the
equivalent of a 2-equation solution if desired.