Allow missing dbz, vel, or sw in radar sweepfile
For optimal data analysis the reflectivity, velocity, and
spectrum width will all be available. This allows for fall
speed calculation and better characterization of the error.
Sometimes its not all there though, for example in WSR-88D
split tilts. This change allows the option to read in data
with only one of the three variables. Further changes to
VarDriver3D.cpp are required to handle the missing values.