Improve boundary condition implementation
Though not that many lines of code, this is a fairly extensive
overhaul of the way BCs were being handled in the analysis.
Previously, the SC filtering was done after the SA transform,
and the BCs were implemented as modified basis functions following
the 'closed' form of the spline transform. In the new implementation,
the BCs are implemented as a matrix transform on the spline
amplitudes themselves following the 'open' form (see Ooyama 2002).
This change necessitated swapping the order of the SC and SA
transforms, so that the filter doesn't destroy the BCs. Likewise,
the FFT has been moved after the SA transform for the same reason. The
change now enforces the BCs exactly and improves numerical precision.
It also paves the way for further optimization since the BCs are no
longer enforced via a complicated switch statement every time the
basis function is called.