Refactor H transform for sparse matrix-vector multiply
In the previous version, the basis functions were re-calculated
at every call of HCq. Since most of that calculation was redundant
the combined basis function and observation weight can be stored
in a sparse matrix that is pre-calculated. This refactor results
in no change to the output but has significant cost savings.