Change adjustment bg error to 1, but allow for configurable value
The background adjustment takes the original bg values and uses
those as pseudo-obs to create a variational solution starting from
the interpolated fields using exponential weighting. The bg error
then is nebulously defined, since it is the same data just with a
different spatial filter. The value was originally 10.0, but tests
indicate a value of 1.0 is good enough, and reduces run time.